feat: make game process use their own uws apps and ports

+++ server/src/game/gameProcessManager.ts
@@ -5,6 +5,7 @@ import { type MapDefKey, MapDefs } from "../../../shared/defs/mapDefs.ts";
 import type { TeamMode } from "../../../shared/gameConfig.ts";
 import * as net from "../../../shared/net/net.ts";
 import { util } from "../../../shared/utils/util.ts";
+import { Config } from "../config.ts";
 import { ServerLogger } from "../utils/logger.ts";
 import { type FindGamePrivateBody, type ServerGameConfig } from "../utils/types.ts";
 import { type GameData, type ProcessMsg, ProcessMsgType } from "./ipcTypes.ts";
@@ -24,6 +25,7 @@ export enum ProcState {
 
 class GameProcess {
     process: ChildProcess;
+    port: number;
 
     gameData: GameData = {
         id: "",
@@ -50,9 +52,16 @@ class GameProcess {
 
     reusedCount = 0;
 
-    constructor(manager: GameProcessManager, id: string, config: ServerGameConfig) {
+    constructor(
+        manager: GameProcessManager,
+        id: string,
+        config: ServerGameConfig,
+        port: number,
+    ) {
         this.manager = manager;
-        this.process = fork(procFile, [], {
+        this.port = port;
+
+        this.process = fork(procFile, [port.toString()], {
             serialization: "advanced",
         });
 
@@ -192,7 +201,17 @@ export class GameProcessManager {
 
     readonly logger = new ServerLogger("Game Process Manager");
 
+    private readonly _freePorts: number[] = [];
+
+    getNextPort() {
+        return this._freePorts.shift();
+    }
+
     constructor() {
+        for (let i = 0; i < Config.gameServer.maxGames; i++) {
+            this._freePorts.push(Config.gameServer.firstGamePort + i);
+        }
+
         process.on("beforeExit", () => {
             for (const gameProc of this.processes) {
                 gameProc.process.kill();
@@ -247,7 +266,7 @@ export class GameProcessManager {
         }, 0);
     }
 
-    newGame(config: ServerGameConfig): GameProcess {
+    newGame(config: ServerGameConfig): GameProcess | undefined {
         let gameProc: GameProcess | undefined;
 
         for (let i = 0; i < this.processes.length; i++) {
@@ -260,13 +279,21 @@ export class GameProcessManager {
 
         const id = randomUUID();
         if (!gameProc) {
-            gameProc = new GameProcess(this, id, config);
+            const port = this.getNextPort();
+            if (port === undefined) {
+                return undefined;
+            }
+            gameProc = new GameProcess(this, id, config, port);
 
             this.processes.push(gameProc);
 
             gameProc.process.on("exit", () => {
                 this.killProcess(gameProc!);
+                if (!this._freePorts.includes(gameProc!.port)) {
+                    this._freePorts.push(gameProc!.port);
+                }
             });
+
             gameProc.process.on("close", () => {
                 this.killProcess(gameProc!);
             });
@@ -308,8 +335,8 @@ export class GameProcessManager {
         return this.processById.get(id);
     }
 
-    async findGame(body: FindGamePrivateBody): Promise<GameProcess> {
-        let proc = this.processes
+    async findGame(body: FindGamePrivateBody): Promise<GameProcess | undefined> {
+        let proc: GameProcess | undefined = this.processes
             .filter((proc) => {
                 const game = proc.gameData;
                 return (
@@ -328,6 +355,10 @@ export class GameProcessManager {
                 teamMode: body.teamMode,
                 mapName: body.mapName as MapDefKey,
             });
+        }
+
+        if (!proc) {
+            return undefined;
         }
 
         // if the game has not finished creating
